update: Sistema de batalha com RNG

+++ inimigo/inimigo.h
@@ -5,8 +5,8 @@ typedef struct inimigo {
     char nome[10];
     int atk;
     int matk;
-    int vida;
-    int vidaMax;
+    float vida;
+    float vidaMax;
     int def;
     int mana;
     int velocidade;
